requirements.txt
duckdb
tqdm
beautifulsoup4
pandas
requests
pyarrow
scikit-learn
numpy
matplotlib
torchruntime
rich
pydantic